Move Netlify env hacks to a separate file
