+++ netlify.toml
@@ -1,4 +1,5 @@
 [build]
+command = "node ./prepare-netlify-dotenv.mjs; yarn build"
 publish = ".next"
 
 ## Yarn 3 cache does not work out of the box as of Jan 2022. Context:
