+++ next.config.mjs
@@ -1,33 +1,6 @@
 // @ts-check
 import { withSentryConfig } from "@sentry/nextjs";
 
-/*
- * Netlify UI does not support context-based environment variables.
- * For example, we cannot configure XYZ=42 for preview releases (pull requests)
- * and XYZ=4242 for production (main branch).
- *
- * As a workaround, we read Netlify’s CONTEXT env variable and then
- * map PROD_XYZ / XYZ_PROD or PR_XYZ / XYZ_PR into XYZ.
- */
-
-const suffixOrPrefixByContext = {
-  production: "PROD",
-  "deploy-preview": "PR",
-};
-
-const suffixOrPrefix = suffixOrPrefixByContext[process.env.CONTEXT];
-
-if (suffixOrPrefix) {
-  for (const key in process.env) {
-    if (key.startsWith(`${suffixOrPrefix}_`)) {
-      process.env[key.slice(suffixOrPrefix.length + 1)] = process.env[key];
-    } else if (key.endsWith(`_${suffixOrPrefix}`)) {
-      process.env[key.slice(undefined, suffixOrPrefix.length - 1)] =
-        process.env[key];
-    }
-  }
-}
-
 // Adding NEXT_PUBLIC_ makes environment variables available inside Next.js client.
 // We don’t use the prefix outside Next.js code to make app setup less bulky and
 // to avoid unnecessary breaking changes in it.
@@ -35,12 +8,6 @@ process.env.NEXT_PUBLIC_COMMENTS_ARE_PAUSED = process.env.COMMENTS_ARE_PAUSED;
 process.env.NEXT_PUBLIC_DJANGO_BASE_URL = process.env.DJANGO_BASE_URL;
 process.env.NEXT_PUBLIC_SENTRY_DSN = process.env.SENTRY_DSN;
 process.env.NEXT_PUBLIC_SENTRY_ENVIRONMENT = process.env.SENTRY_ENVIRONMENT;
-
-// Configure Auth0 base URL based on Netlify env. This enables auth in deploy previews.
-// https://docs.netlify.com/configure-builds/environment-variables/#deploy-urls-and-metadata
-if (!process.env.AUTH0_BASE_URL) {
-  process.env.AUTH0_BASE_URL = process.env.DEPLOY_PRIME_URL;
-}
 
 /**
  * @type Omit<import("next").NextConfig, "webpack">
